+++ angular-workspace/projects/example-client-app/src/app/app.module.ts
@@ -11,6 +11,7 @@ import { NimbleTextAreaModule, NimbleTextFieldModule, NimbleNumberFieldModule, N
     NimbleAnchorModule, NimbleAnchorButtonModule, NimbleAnchorTabModule, NimbleAnchorTabsModule,
     NimbleIconCheckModule, NimbleBannerModule, NimbleAnchorMenuItemModule, NimbleAnchorTreeItemModule } from '@ni/nimble-angular';
 import { NimbleLabelProviderCoreModule } from '@ni/nimble-angular/label-provider/core';
+import { NimbleLabelProviderRichTextModule } from '@ni/nimble-angular/label-provider/rich-text';
 import { NimbleLabelProviderTableModule } from '@ni/nimble-angular/label-provider/table';
 import { NimbleMappingTextModule } from '@ni/nimble-angular/mapping/text';
 import { NimbleMappingIconModule } from '@ni/nimble-angular/mapping/icon';
@@ -22,8 +23,8 @@ import { NimbleTableColumnDateTextModule } from '@ni/nimble-angular/table-column
 import { NimbleTableColumnEnumTextModule } from '@ni/nimble-angular/table-column/enum-text';
 import { NimbleTableColumnIconModule } from '@ni/nimble-angular/table-column/icon';
 import { NimbleTableColumnNumberTextModule } from '@ni/nimble-angular/table-column/number-text';
-import { NimbleRichTextViewerModule } from '@ni/nimble-angular/rich-text-viewer';
-import { NimbleRichTextEditorModule } from '@ni/nimble-angular/rich-text-editor';
+import { NimbleRichTextViewerModule } from '@ni/nimble-angular/rich-text/viewer';
+import { NimbleRichTextEditorModule } from '@ni/nimble-angular/rich-text/editor';
 import { AppComponent } from './app.component';
 import { CustomAppComponent } from './customapp/customapp.component';
 import { HeaderComponent } from './header/header.component';
@@ -40,6 +41,7 @@ import { HeaderComponent } from './header/header.component';
         ReactiveFormsModule,
         NimbleThemeProviderModule,
         NimbleLabelProviderCoreModule,
+        NimbleLabelProviderRichTextModule,
         NimbleLabelProviderTableModule,
         NimbleTextAreaModule,
         NimbleTextFieldModule,