+++ packages/nimble-components/src/table-column/number-text/index.ts
@@ -1,4 +1,7 @@
-import { DesignSystem } from '@microsoft/fast-foundation';
+import {
+    DesignSystem,
+    DesignTokenSubscriber
+} from '@microsoft/fast-foundation';
 import { attr, nullableNumberConverter } from '@microsoft/fast-element';
 import { styles } from '../base/styles';
 import { template } from '../base/template';
@@ -10,11 +13,11 @@ import { tableColumnNumberTextCellViewTag } from './cell-view';
 import type { ColumnInternalsOptions } from '../base/models/column-internals';
 import { NumberTextAlignment, NumberTextFormat } from './types';
 import type { NumberFormatter } from './models/number-formatter';
-import { RoundToIntegerFormatter } from './models/round-to-integer-formatter';
 import { DefaultFormatter } from './models/default-formatter';
 import { DecimalFormatter } from './models/decimal-formatter';
 import { TableColumnNumberTextValidator } from './models/table-column-number-text-validator';
 import { TextCellViewBaseAlignment } from '../text-base/cell-view/types';
+import { lang } from '../../theme-provider';
 
 export type TableColumnNumberTextCellRecord = TableNumberField<'value'>;
 export interface TableColumnNumberTextColumnConfig {
@@ -46,9 +49,21 @@ export class TableColumnNumberText extends TableColumnTextBase {
     @attr({ attribute: 'decimal-digits', converter: nullableNumberConverter })
     public decimalDigits?: number;
 
+    private readonly langSubscriber: DesignTokenSubscriber<typeof lang> = {
+        handleChange: () => {
+            this.updateColumnConfig();
+        }
+    };
+
     public override connectedCallback(): void {
         super.connectedCallback();
+        lang.subscribe(this.langSubscriber, this);
         this.updateColumnConfig();
+    }
+
+    public override disconnectedCallback(): void {
+        super.disconnectedCallback();
+        lang.unsubscribe(this.langSubscriber, this);
     }
 
     public override get validity(): TableColumnValidity {
@@ -93,14 +108,13 @@ export class TableColumnNumberText extends TableColumnTextBase {
 
     private createFormatter(): NumberFormatter {
         switch (this.format) {
-            case NumberTextFormat.roundToInteger:
-                return new RoundToIntegerFormatter();
             case NumberTextFormat.decimal:
                 return new DecimalFormatter(
+                    lang.getValueFor(this),
                     this.decimalDigits ?? defaultDecimalDigits
                 );
             default:
-                return new DefaultFormatter();
+                return new DefaultFormatter(lang.getValueFor(this));
         }
     }
 
@@ -114,10 +128,7 @@ export class TableColumnNumberText extends TableColumnTextBase {
         }
 
         // Look at format to determine the default alignment
-        if (
-            this.format === NumberTextFormat.roundToInteger
-            || this.format === NumberTextFormat.decimal
-        ) {
+        if (this.format === NumberTextFormat.decimal) {
             return TextCellViewBaseAlignment.right;
         }
         return TextCellViewBaseAlignment.left;