+++ packages/nimble-components/src/table-column/number-text/tests/table-column-number-text.stories.ts
@@ -9,6 +9,7 @@ import { tableTag } from '../../../table';
 import { tableColumnNumberTextTag } from '..';
 import {
     SharedTableArgs,
+    columnOperationBehavior,
     sharedTableActions,
     sharedTableArgTypes,
     sharedTableArgs
@@ -49,20 +50,11 @@ const simpleData = [
     }
 ];
 
-const overviewText = `This page contains information about the types of columns that can be displayed in a \`nimble-table\`.
-See the **Table** page for information about configuring the table itself and the **Table Column Configuration** page for
-information about common column configuration.`;
-
 const metadata: Meta<SharedTableArgs> = {
     title: 'Incubating/Table Column - Number Text',
     decorators: [withActions],
     tags: ['autodocs'],
     parameters: {
-        docs: {
-            description: {
-                component: overviewText
-            }
-        },
         actions: {
             handles: sharedTableActions
         }
@@ -92,8 +84,10 @@ interface NumberTextColumnTableArgs extends SharedTableArgs {
     validity: () => void;
 }
 
-const numberTextColumnDescription = `The \`nimble-table-column-number-text\` column is used to display number fields as text in the \`nimble-table\`. Column operations, such as sorting and grouping,
-are performed numerically on the data values.`;
+const numberTextColumnDescription = `The \`nimble-table-column-number-text\` column is used to display number fields as text in the \`nimble-table\`. Numbers are formatted in a locale-specific way
+based on the value of the \`lang\` token, which can be set via the [\`nimble-theme-provider\`](?path=/docs/tokens-theme-provider--docs).
+
+${columnOperationBehavior}`;
 
 const formatDescription = `Configures the way that the numeric value is formatted to render within the column.
 
@@ -103,9 +97,8 @@ const formatDescription = `Configures the way that the numeric value is formatte
     <ul>
         <li>\`default\`: Integers are shown with no trailing zeros, the value is limited to 6 digits, and exponential notation is used for numbers that are large (\`>= 1e6\`) or small (\`< 1e-3\`) in magnitude.
         </li>
-        <li>\`decimal\`: Values as are formatted as decimal values, always displaying \`decimal-digits\` digits after the separator and never displaying exponential notation.
-        </li>
-        <li>\`roundToInteger\`: Values are rounded to the nearest whole number. Exponential notation is never used. It can only safely represent integers up to the magnitude of \`MAX_SAFE_INTEGER\`.
+        <li>\`decimal\`: Values as are formatted as decimal values, always displaying \`decimal-digits\` digits after the separator and never displaying exponential notation. Setting \`decimal-digits\` to \`0\`
+        will display the value as an integer without a decimal separator.
         </li>
     </ul>
 </details>
@@ -128,8 +121,6 @@ To improve the ability for users to visually scan values, applications should se
         <li>\`default\` format: Values are left-aligned.
         </li>
         <li>\`decimal\` format: Values are right-aligned.
-        </li>
-        <li>\`roundToInteger\` format: Values are right-aligned.
         </li>
     </ul>
 </details>